backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
httpx==0.27.2
google-api-python-client==2.152.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googlemaps==4.10.0
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
pymongo==4.6.1
pytest==8.3.3
ruff==0.8.6
